+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * improved and fixed the `AuthHelper`. Now it is loaded from the view (instead of
     from the controller) requires an array with the `user` key as configuration;
 * improved validation rules. Removed some useless validation methods;
+* added `MeCms\AuthTrait`. It provides some methods for classes that need to
+    verify the data of the logged in user;
 * added `MeCms\Validation\BannerAndPhotoValidator` abstract class;
 * added `MeCms\Validation\PageAndPostValidator` abstract class.